Add config timelines.shouldSyncInitially
Prevent 100 tweets per timeline being polled for on startup if this is set to false. Default to true for backwards compat.

If set to false, the first poll for a timeline (on startup `tline.hasProcessedTweets = false`, on provision or DB deletion, since = null) will be done with `req.count = 1`, so that only one tweet is received. This tweet is the most recent tweet and is not bridged, but it's `since` is stored so that future polls will be restricted to tweets that were created after startup.